initialize repo
This is a basic version, which would check the website of the
swimming school periodically and then if some patterns are not found
on the website, like "No Single Lessons", it will send a
notification.

This is really raw, but this might be helpful. The thing that will
need to be considered is how to not get spammed when there will be
some availability.
But my brain is off today as I'm totally under the weather, I will
check that another day...