fix empty content != no previous content
To know if we were having a previous content we were checking if the
string was a zero value (empty string). The problem is that from now
on the school is putting an image when there is no availability,
ending up with a content being empty after parsing. In consequence
the application was thinking that no previous content was fetched
before.

So from now on, we have the content, a boolean and an error as a
result, like that we can differentiate not found and empty